// Healty Meal Service/Healthy-Meals-Service-Frontend/src/app/components/partner/model/Partner.ts
export interface Partner {
    partnerId: number;
    partnerName: string;
    shopName: string;
    partnerEmail: string;
    phoneNumber: string;
    password: string;
    createdAt: string;
    address: string;
    city: string;
    state: string;
    zipCode: string;
    shopDescription?: string;
    logoUrl?: string;
  }
  